requirements.txt
matplotlib
tensorflow=1.10.0
pytorch
bert-serving-server
bert-serving-client
torchtext
conllu